requirements.txt
cffi==1.14.4
cryptography==3.3.2
numpy==1.23.5
pandas==1.5.1
pycparser==2.20
PyMySQL==0.10.1
python-dateutil==2.8.2
python-dotenv==0.15.0
pytz==2022.6
six==1.15.0
